requirements.txt
PySide6>=6.5.0
pyserial>=3.5
PyOpenGL>=3.1.7
numpy>=1.24.0
